+++ src/commands/setupSourceDirectory.ts
@@ -1,4 +1,6 @@
 import * as vscode from 'vscode';
+import { posix } from 'path';
+import { MACRO_EXTENSION } from '../core/constants';
 import { ExtensionContext } from '../extensionContext';
 import { selectSourceDirectory } from '../ui/selectMacroFile';
 import { readFile } from '../utils/resources';
@@ -7,7 +9,24 @@ import { PathLike, toUri } from '../utils/uri';
 export const GLOBALS_RESOURCE = 'api/global.d.ts';
 export const JSCONFIG_RESOURCE = 'api/jsconfig.json';
 
-export async function setupSourceDirectory(context: ExtensionContext, pathOrUri?: PathLike): Promise<void> {
+export function registerSourceDirectoryVerifier(context: ExtensionContext) {
+  const verifiedDir = new Set<string>();
+  return vscode.workspace.onDidSaveTextDocument(async ({ uri }) => {
+    if (!uri.path.endsWith(MACRO_EXTENSION)) {
+      return;
+    }
+
+    const parentPath = posix.dirname(uri.path);
+    if (verifiedDir.has(parentPath)) {
+      return;
+    }
+
+    verifiedDir.add(parentPath);
+    await setupSourceDirectory(context, uri.with({ path: parentPath }), true);
+  });
+}
+
+export async function setupSourceDirectory(context: ExtensionContext, pathOrUri?: PathLike, suppressNotifications?: true): Promise<void> {
   const uri = pathOrUri ? toUri(pathOrUri) : await selectSourceDirectory(context.libraryManager);
   if (!uri) {
     return; // Nothing to run.
@@ -21,20 +40,25 @@ export async function setupSourceDirectory(context: ExtensionContext, pathOrUri?
   ])).some((result) => result);
 
   if (!updatingFiles) {
-    vscode.window.showInformationMessage('All files are up-to-date.');
+    if (!suppressNotifications) {
+      vscode.window.showInformationMessage('All files are up-to-date.');
+    }
     return; // Nothing to update.
   }
 
-  if (await vscode.workspace.applyEdit(edit)) {
-    vscode.window.showInformationMessage('Updated files to the latest versions.');
-  } else {
-    vscode.window.showErrorMessage(`Could not update ${uri.fsPath}`);
+  const result = await vscode.workspace.applyEdit(edit);
+  if (!suppressNotifications) {
+    if (result) {
+      vscode.window.showInformationMessage('Updated files to the latest versions.');
+    } else {
+      vscode.window.showErrorMessage(`Could not update ${uri.fsPath}`);
+    }
   }
 
   async function update(uri: vscode.Uri, source: string, target: string): Promise<boolean> {
     const [currentContents, newContents] = await Promise.all([
       vscode.workspace.openTextDocument(vscode.Uri.joinPath(uri, target))
-        .then((d) => d.getText(), () => undefined),
+        .then((d) => d.getText(), () => ''),
       readFile(context.extensionContext, source),
     ]);
     if (currentContents === newContents) {